Decouple Search in a new class
